Basal Plan - Add history page (#6)
* Refactored the database code into a static `BasalDB` class.

* Moved `basal_plan_home_page.dart` to a dedicated `pages` folder.

* Added `totalBasal` property to `BasalPlan`.

* Added a `plans` method to `BasalDB` to get an (ordered) `Stream` of plans.

* Added a basic chart widget that can show a `BasalPlan` in a visual way.

* Added a basic history page that shows the overviews of previous plans.

The page currently is read-only and does not yet support deleting
previous plans or viewing a specific one.

* Added a `BasalPlanOverviewTile` widget.
* Added a basic `BasalPlanHistoryPage` page.

* Updated `BasalPlanHistoryPage` to not show the current plan.

* Added `onTap` and `onLongPress` to `BasalPlanOverviewTile`.

* Added `HidingProgressIndicator` to `BasalPlanHistoryPage` to show only
when data is received.

* Added `removePlan` to `BasalDB`.

* Created a `_BasalPlanIdx` opaque type that adds an attached
`DocumentReference` to a `BasalPlan`. This enables us to hide the
complexity from the user while keeping track of this information, which
is needed for removing a document. The caveat here is that this won't
work with plans created by the user.

* Added a `key` to the `BasalPlanHistoryPage` `ListView` for more
responsive adding / removing.

* Added option to delete a plan from the history.

* Added `TileChosenBottomSheet` - a bottom sheet that contains buttons
to delete and show a plan.

* Moved `DateTimeExtensions` from `basal_plan_overview_tile.dart` to a
dedicated file, to enable code sharing.

* Fixed exception in `BasalSegmentTile`.

The exception occoured when we used a `Dismissible` with `direction ==
null`. To solve this, we only wrap the tile in a dismissible when we
want it to be enabled.

* Removed implicit `Navigator.pop` call in the `TileChosenBottomSheet`
callbacks.

The reason for the removal was to enable a `Navigator.Push` call to move
to another page. The previous behaviour immediately called `pop`, and made
it impossible to navigate to a new page.

* Added costumization to the `BasalPlanChart`.

* Line color
* Grid color
* Title text style

* Added `BasalPlanPage`, which shows an "overview" of a specific plan.

The page shows a chart of the plan and then a listview of the segments
in readonly mode.

* Implemented the `onShowSelected` callback in `BasalPlanHistoryPage` to
show the `BasalPlanPage`.

* Added a `MultiToolPage` interface.

* Changed the ratio between the chart and the listview in `BasalPlanPage`.

* Fixed animation bug in the FAB of `BasalPlanHomePage`.

To prevent the animation from happening, the FAB needs to always be part
of the widget tree, so instead of using a ternary operator, it was
changed to use a `Visibility` widget.

* Removed the `AppBar` from the pages, and created a unified `Scaffold`
with a `Drawer` in `MyApp` that allows selecting the different pages.